--- a/src/BlazorDatasheet.Core/Data/Sheet.cs
+++ b/src/BlazorDatasheet.Core/Data/Sheet.cs
@@ -336,7 +336,9 @@ public class Sheet
     /// <param name="col"></param>
     internal void MarkDirty(int row, int col)
     {
-        MarkDirty(new Region(row, col));
+        if (!Region.Contains(row, col))
+            return;
+        _dirtyRows.Set(row, row, true);
     }
 
     /// <summary>
@@ -345,7 +347,7 @@ public class Sheet
     /// <param name="region"></param>
     internal void MarkDirty(IRegion region)
     {
-        if(!region.Intersects(this.Region))
+        if (!region.Intersects(Region))
             return;
 
         _dirtyRows.Set(region.Top, region.Bottom, true);
@@ -361,7 +363,7 @@ public class Sheet
     {
         foreach (var region in regions)
         {
-            if(!region.Intersects(this.Region))
+            if (!region.Intersects(this.Region))
                 continue;
 
             _dirtyRows.Set(region.Top, region.Bottom, true);