--- a/src/BlazorDatasheet.DataStructures/Store/MatrixRestoreData.cs
+++ b/src/BlazorDatasheet.DataStructures/Store/MatrixRestoreData.cs
@@ -3,11 +3,17 @@ namespace BlazorDatasheet.DataStructures.Store;
 public class MatrixRestoreData<T>
 {
     public List<(int row, int col, T? data)> DataRemoved { get; internal set; } = new();
-    public List<AppliedShift> Shifts { get; internal set; } = new();
+    public List<AppliedShift>? Shifts { get; internal set; }
 
     public void Merge(MatrixRestoreData<T> item)
     {
         DataRemoved.AddRange(item.DataRemoved);
-        Shifts.AddRange(item.Shifts);
+        if (item.Shifts != null)
+        {
+            if (Shifts == null)
+                Shifts = item.Shifts;
+            else
+                Shifts.AddRange(item.Shifts);
+        }
     }
 }