--- a/src/BlazorDatasheet.DataStructures/Store/RegionRestoreData.cs
+++ b/src/BlazorDatasheet.DataStructures/Store/RegionRestoreData.cs
@@ -6,13 +6,21 @@ public class RegionRestoreData<T>
 {
     public List<DataRegion<T>> RegionsRemoved { get; init; } = new();
     public List<DataRegion<T>> RegionsAdded { get; init; } = new();
-    public List<AppliedShift> Shifts { get; init; } = new();
+    public List<AppliedShift>? Shifts { get; set; } = null;
 
     public RegionRestoreData<T> Merge(RegionRestoreData<T> item)
     {
         RegionsAdded.AddRange(item.RegionsAdded);
         RegionsRemoved.AddRange(item.RegionsRemoved);
-        Shifts.AddRange(item.Shifts);
+
+        if (item.Shifts != null)
+        {
+            if (Shifts == null)
+                Shifts = new List<AppliedShift>(item.Shifts);
+            else
+                Shifts.AddRange(item.Shifts);
+        }
+
         return this;
     }
 }