--- a/src/BlazorDatasheet.Formula.Core/Dependencies/VertexType.cs
+++ b/src/BlazorDatasheet.Formula.Core/Dependencies/VertexType.cs
@@ -3,6 +3,5 @@
 public enum VertexType
 {
     Cell,
-    Region,
     Named
 }