--- a/test/BlazorDatasheet.Test/Formula/CellReferenceTests.cs
+++ b/test/BlazorDatasheet.Test/Formula/CellReferenceTests.cs
@@ -22,6 +22,7 @@ public class CellReferenceTests
     [TestCase("$A$1", 0, 0, true, true)]
     [TestCase("AA102", 101, 26, false, false)]
     [TestCase("AAA102", 101, 702, false, false)]
+    [TestCase("$NTP$1", 0, 9_999, true, true)]
     public void Cell_Str_Parses_Correctly(string cellStr, int rowExpected, int colExpected, bool rowAbsExpected,
         bool colAbsExpected)
     {