Picks some low-hanging fruit for improving formula performance. Still uses a lot of memory for setting formula, mostly in the dependency tracking and for storing undo data.

Following function was used to assess performance.
```csharp
_sheet.BatchUpdates();

for (int i = 0; i < 200; i++)
{
    for (int j = 0; j < 200; j++)
    {
        var cell = RangeText.ToCellText(i + 1, j);
        _sheet.Cells.SetFormula(i, j, $"={cell}+1");
    }
}

_sheet.EndBatchUpdates();
```


Before this PR:
| Method         | Mean     | Error    | StdDev   | Allocated |
|--------------- |---------:|---------:|---------:|----------:|
| Set200x200Formula | 709.8 ms | 20.09 ms | 57.32 ms | 889.48 MB |

After this PR:
| Method         | Mean     | Error    | StdDev  | Allocated |
|--------------- |---------:|---------:|--------:|----------:|
| Set200x200Formula | 574.7 ms | 10.72 ms | 9.50 ms | 657.94 MB |